+++ tapis-appsapi/src/main/java/edu/utexas/tacc/tapis/apps/api/resources/AppResource.java
@@ -1105,7 +1105,7 @@ public class AppResource
     var jobAttrs = req.jobAttributes;
     if (jobAttrs == null) jobAttrs = new JobAttributes();
     // Extract Notes from the raw json.
-    Object notes = extractNotes(rawJson);
+    JsonObject notes = extractNotes(rawJson);
     // Create App
     var app = new App(-1, -1, tenantId, req.id, req.version, req.description, req.jobType, req.owner, req.enabled,
           DEFAULT_CONTAINERIZED,  req.runtime, req.runtimeVersion, req.runtimeOptions, req.containerImage,
@@ -1128,7 +1128,7 @@ public class AppResource
     var jobAttrs = req.jobAttributes;
     if (jobAttrs == null) jobAttrs = new JobAttributes();
     // Extract Notes from the raw json.
-    Object notes = extractNotes(rawJson);
+    JsonObject notes = extractNotes(rawJson);
 
     // NOTE: Following attributes are not updatable and must be filled in on service side.
     String owner = null;
@@ -1179,9 +1179,9 @@ public class AppResource
    * approaches caused problems with the json marshalling. This method ensures notes end up as a JsonObject rather
    * than a LinkedTreeMap.
    */
-  private static Object extractNotes(String rawJson)
+  private static JsonObject extractNotes(String rawJson)
   {
-    Object notes = null;
+    JsonObject notes = null;
     // Check inputs
     if (StringUtils.isBlank(rawJson)) return notes;
     // Turn the request string into a json object and extract the notes object